refactor
Adjust test structure to match new crate structure.
